Created NostrConfig, Beans conditional on property nostr.active
