+++ docker-compose.yml
@@ -87,6 +87,7 @@ services:
       - ADMIN_MACAROON_PATH=${ADMIN_MACAROON_PATH}
       - LND_TLS_CERT_PATH=${LND_TLS_CERT_PATH}
       - KAFKA_BOOTSTRAP_SERVERS=${KAFKA_BOOTSTRAP_SERVERS}
+      - NOSTR_ACTIVE=${NOSTR_ACTIVE:-true}
       - NOSTR_RELAY_URL=${NOSTR_RELAY_URL:-wss://relay.primal.net}
       - NOSTR_RETRY_MAX=${NOSTR_RETRY_MAX:-5}
       - NOSTR_RETRY_INITIAL_DELAY=${NOSTR_RETRY_INITIAL_DELAY:-2000}
