+++ src/main/java/com/aratiri/aratiri/nostr/NostrConfig.java
@@ -1,0 +1,33 @@
+package com.aratiri.aratiri.nostr;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class NostrConfig {
+
+    @Bean
+    public NostrClient nostrClient(@Value("${nostr.active:false}") boolean active) {
+        if (!active) {
+            return new NostrClientDisabled();
+        }
+        return new NostrClientImpl();
+    }
+
+    @Bean
+    public NostrService nostrService(
+            NostrClient nostrClient,
+            RestTemplate restTemplate,
+            ObjectMapper objectMapper,
+            @Value("${nostr.active:false}") boolean active) {
+        if (active) {
+            return new NostrServiceImpl(restTemplate, nostrClient, objectMapper);
+        } else {
+            return new NostrServiceDisabled();
+        }
+    }
+
+}
